Agrego tema puntos y expiracion a la reserva.
Falta una cosa que no se actualizan bien los puntos cuando se muewstran en pantalla.

El tema de las fechas es una cagada por ahora subo eso que se visualiza mal. El timezone es una mierda, puto ISO.

+++ trunk/src/main/java/com/canchita/DAO/ExpirationDAO.java
@@ -16,9 +16,9 @@ public interface ExpirationDAO {
 
 	public Expiration getById(Long id) throws PersistenceException;
 
-	public Expiration getByScore(Field aField, Integer score) throws ElementNotExistsException;
+	public Expiration getByScore(Field aField, Long score) throws ElementNotExistsException;
 	
-	public Expiration getByScore(Complex aComplex, Integer score) throws ElementNotExistsException;
+	public Expiration getByScore(Complex aComplex, Long score) throws ElementNotExistsException;
 	
 	public void update(Expiration expiration) throws PersistenceException;
